Split ActionManager.py into ActionManagerParts/
